+++ eqsig/__about__.py
@@ -1,4 +1,4 @@
 __project__ = "eqsig"
 __author__ = "Maxim Millen"
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 __license__ = "MIT"